fix: normalize AI student identifiers

+++ EduOS.Core/Entities/AI/AttendancePrediction.cs
@@ -5,7 +5,7 @@ namespace EduOS.Core.Entities.AI
 {
     public class AttendancePrediction : BaseTenantEntity
     {
-        public int StudentId { get; set; }
+        public long StudentId { get; set; }
         public DateTime PredictionDate { get; set; }
         public decimal ProbabilityOfAbsence { get; set; }
         public string SuggestedAction { get; set; }
